Prepend newline in index string.
This is because of a package.el bug where it searches for the contents,
which is at the end of the headers. Without the newline, and end-of-file
error occurs.

This does not seem to happen all the time and I have not found a way to
continuously reproduce it. Though I have seen that it behaves
differently on various web servers.